Move classes to `support.scanning` package
