+++ junit-platform-commons/src/main/java/org/junit/platform/commons/support/scanning/ClasspathFilters.java
@@ -8,7 +8,7 @@
  * https://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junit.platform.commons.util;
+package org.junit.platform.commons.support.scanning;
 
 import java.nio.file.Path;
 import java.util.function.Predicate;
