+++ junit-platform-commons/src/main/java/org/junit/platform/commons/util/ClasspathScannerLoader.java
@@ -18,6 +18,7 @@ import java.util.ServiceLoader;
 
 import org.junit.platform.commons.JUnitException;
 import org.junit.platform.commons.support.scanning.ClasspathScanner;
+import org.junit.platform.commons.support.scanning.DefaultClasspathScanner;
 
 /**
  * @since 1.12
