+++ junit-platform-commons/src/main/java/org/junit/platform/commons/util/CloseablePath.java
@@ -11,6 +11,7 @@
 package org.junit.platform.commons.util;
 
 import static java.util.Collections.emptyMap;
+import static org.apiguardian.api.API.Status.INTERNAL;
 
 import java.io.Closeable;
 import java.io.IOException;
@@ -27,10 +28,19 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 
+import org.apiguardian.api.API;
+
 /**
+ * <h2>DISCLAIMER</h2>
+ *
+ * <p>These utilities are intended solely for usage within the JUnit framework
+ * itself. <strong>Any usage by external parties is not supported.</strong>
+ * Use at your own risk!
+ *
  * @since 1.0
  */
-final class CloseablePath implements Closeable {
+@API(status = INTERNAL, since = "1.12")
+public final class CloseablePath implements Closeable {
 
 	private static final String FILE_URI_SCHEME = "file";
 	static final String JAR_URI_SCHEME = "jar";
@@ -47,7 +57,7 @@ final class CloseablePath implements Closeable {
 	private final Path path;
 	private final Closeable delegate;
 
-	static CloseablePath create(URI uri) throws URISyntaxException {
+	public static CloseablePath create(URI uri) throws URISyntaxException {
 		return create(uri, it -> FileSystems.newFileSystem(it, emptyMap()));
 	}
